JAR packaging Changed package name to match artifact group id
