+++ Translator/pom.xml
@@ -4,13 +4,132 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.example</groupId>
-    <artifactId>Translator</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <packaging>jar</packaging>
+    <groupId>io.github.hephaestus-metrics</groupId>
+    <artifactId>translator</artifactId>
+    <version>0.1-SNAPSHOT</version>
+
+    <name>Hephaestus metrics</name>
+    <description>A library for accessing metrics collected by Prometheus</description>
+    <url>https://github.com/Hephaestus-Metrics/Translator</url>
+
+    <licenses>
+        <license>
+            <name>MIT License</name>
+            <url>https://www.opensource.org/licenses/mit-license.php</url>
+        </license>
+    </licenses>
+
+    <developers>
+        <developer>
+            <name>Wojciech Piechaczek</name>
+            <email>wojtek.piechaczek@gmail.com</email>
+            <organization>com.github.TheVoidek</organization>
+            <organizationUrl>https://github.com/TheVoidek/</organizationUrl>
+        </developer>
+        <developer>
+            <name>Kinga Sąkól</name>
+            <email>kinga.sakol@wp.pl</email>
+            <organization>com.github.kingasakol</organization>
+            <organizationUrl>https://github.com/kingasakol/</organizationUrl>
+        </developer>
+        <developer>
+            <name>Michał Szczurek</name>
+            <email>michalszczurek@o2.pl</email>
+            <organization>com.github.Ra77us</organization>
+            <organizationUrl>https://github.com/Ra77us/</organizationUrl>
+        </developer>
+        <developer>
+            <name>Krystian Walat</name>
+            <email>krystian.walat@gmail.com</email>
+            <organization>com.github.Dewkong</organization>
+            <organizationUrl>https://github.com/Dewkong/</organizationUrl>
+        </developer>
+    </developers>
+
+    <scm>
+        <connection>scm:git:git://github.com/Hephaestus-Metrics/Translator.git</connection>
+        <developerConnection>scm:git:ssh://github.com:Hephaestus-Metrics/Translator.git</developerConnection>
+        <url>http://github.com/Hephaestus-Metrics/Translator/tree/main</url>
+    </scm>
 
     <properties>
         <maven.compiler.source>16</maven.compiler.source>
         <maven.compiler.target>16</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
+
+    <profiles>
+        <profile>
+            <id>release</id>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-source-plugin</artifactId>
+                        <version>2.2.1</version>
+                        <executions>
+                            <execution>
+                                <id>attach-sources</id>
+                                <goals>
+                                    <goal>jar-no-fork</goal>
+                                </goals>
+                            </execution>
+                        </executions>
+                    </plugin>
+<!--                    Java doc plugin is temporarily disabled while we don't have javadocs -->
+<!--                    <plugin>-->
+<!--                        <groupId>org.apache.maven.plugins</groupId>-->
+<!--                        <artifactId>maven-javadoc-plugin</artifactId>-->
+<!--                        <version>2.9.1</version>-->
+<!--                        <executions>-->
+<!--                            <execution>-->
+<!--                                <id>attach-javadocs</id>-->
+<!--                                <goals>-->
+<!--                                    <goal>jar</goal>-->
+<!--                                </goals>-->
+<!--                            </execution>-->
+<!--                        </executions>-->
+<!--                    </plugin>-->
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-gpg-plugin</artifactId>
+                        <version>1.5</version>
+                        <executions>
+                            <execution>
+                                <id>sign-artifacts</id>
+                                <phase>verify</phase>
+                                <goals>
+                                    <goal>sign</goal>
+                                </goals>
+                            </execution>
+                        </executions>
+                    </plugin>
+                    <plugin>
+                        <groupId>org.sonatype.plugins</groupId>
+                        <artifactId>nexus-staging-maven-plugin</artifactId>
+                        <version>1.6.7</version>
+                        <extensions>true</extensions>
+                        <configuration>
+                            <serverId>ossrh</serverId>
+                            <nexusUrl>https://s01.oss.sonatype.org/</nexusUrl>
+                            <autoReleaseAfterClose>true</autoReleaseAfterClose>
+                        </configuration>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+    </profiles>
+
+    <distributionManagement>
+        <snapshotRepository>
+            <id>ossrh</id>
+            <url>https://s01.oss.sonatype.org/content/repositories/snapshots</url>
+        </snapshotRepository>
+        <repository>
+            <id>ossrh</id>
+            <url>https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/</url>
+        </repository>
+    </distributionManagement>
 
 </project>
