+++ Translator/src/main/java/io/github/hephaestusmetrics/model/QueryInfo.java
@@ -1,4 +1,4 @@
-package main.model;
+package io.github.hephaestusmetrics.model;
 
 import java.util.List;
