+++ Translator/src/main/java/io/github/hephaestusmetrics/model/utilities/ValueConverter.java
@@ -1,4 +1,4 @@
-package main.model.utilities;
+package io.github.hephaestusmetrics.model.utilities;
 
 /**
  * Converts [ <unix_time>, "<sample_value>" ] from string to doubles
